Rename: and relocate: `ColonyDomainSelector`
