+++ src/modules/dashboard/components/ColonyHome/ColonyDomainSelector/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ColonyDomainSelector';
