--- src/modules/dashboard/components/DomainDropdown/index.ts
@@ -1,1 +1,0 @@
-export { default } from './DomainDropdown';
